fixed gui bug
fixed the stupid gui bug, was caused by a getch() that i had removed. dunno
why the getch() made a difference, but i put it back in, hidden inside the
BoardGUI constructor. also made a smart BoardGUI constructor inside
GUI_Globals. main is thus a bit simplified

also fixed a bug that was re-introduced in new_game, where the last piece
to be inserted caused a @ to be drawn on the gui to the right of the cursor